Updates AST css and adds vanilla css for media
I did some more Google searching after the previous commit and found a generator
for the AST version of the CSS. I went ahead and created a vanilla CSS version
for future use and included it in the project. I also updated the mistakes I
had made when trying to make the AST version of the CSS by hand. This includes
those fixes as well.